@@ -1,5 +1,16 @@
+import { createReadStream } from 'fs';
+import path from 'path';
+
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
 const read = async () => {
-    // Write your code here 
+    const filePath = path.join(__dirname, 'files', 'fileToRead.txt');
+    const readStream = createReadStream(filePath);
+  
+    readStream.pipe(process.stdout);
 };
 
 await read();
