@@ -1,5 +1,16 @@
+import { createWriteStream } from 'fs';
+import path from 'path';
+
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
 const write = async () => {
-    // Write your code here 
+    const filePath = path.join(__dirname, 'files', 'fileToWrite.txt');
+    const writeStream = createWriteStream(filePath);
+  
+    process.stdin.pipe(writeStream);
 };
 
 await write();
